refactor: Move to appwrite.py
